+++ Capstone-Api/Capstone-Api/Models/Message.cs
@@ -20,9 +20,9 @@ namespace Capstone_Api.Models
         public DateTime Date { get; set; }
 
         [Required]
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
         [Required]
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
     }
 }
